リファクタリング: クラスを切り出し
